+++ apps/nouns-camp/src/app/candidates/[id]/page.js
@@ -1,4 +1,5 @@
 import { headers } from "next/headers";
+import { notFound as nextNotFound } from "next/navigation";
 import {
   string as stringUtils,
   markdown as markdownUtils,
@@ -9,9 +10,16 @@ import { getStateFromCookie as getWagmiStateFromCookie } from "../../../wagmi-co
 import { subgraphFetch, parseCandidate } from "../../../nouns-subgraph.js";
 import { normalizeId } from "../../../utils/candidates.js";
 import { mainnet } from "../../../chains.js";
-import CandidateScreenClientWrapper from "./page.client.js";
+import { Hydrater as StoreHydrater } from "../../../store.js";
+import ClientAppProvider from "../../client-app-provider.js";
+import CandidateScreen from "../../../components/proposal-candidate-screen.js";
 
 export const runtime = "edge";
+
+const getChainId = () => {
+  const wagmiState = getWagmiStateFromCookie(headers().get("cookie"));
+  return wagmiState?.chainId ?? mainnet.id;
+};
 
 const fetchCandidate = async (id, { chainId }) => {
   const data = await subgraphFetch({
@@ -20,9 +28,20 @@ const fetchCandidate = async (id, { chainId }) => {
       query {
         proposalCandidate(id: "${id}") {
           id
+          slug
+          proposer
+          createdBlock
+          canceledBlock
+          lastUpdatedBlock
+          canceledTimestamp
+          createdTimestamp
+          lastUpdatedTimestamp
           latestVersion {
+            id
             content {
               description
+              matchingProposalIds
+              proposalIdToUpdate
             }
           }
         }
@@ -33,16 +52,16 @@ const fetchCandidate = async (id, { chainId }) => {
 
   return parseCandidate(data.proposalCandidate, { chainId });
 };
+const parseId = (id) => normalizeId(decodeURIComponent(id));
 
 export async function generateMetadata({ params }) {
-  const wagmiState = getWagmiStateFromCookie(headers().get("cookie"));
-  const candidateId = normalizeId(decodeURIComponent(params.id));
+  const candidateId = parseId(params.id);
   const candidate = await fetchCandidate(candidateId, {
-    chainId: wagmiState?.chainId ?? mainnet.id,
+    chainId: getChainId(),
   });
 
   // Can’t notFound() here since we might be on a testnet
-  if (candidate == null) return null;
+  if (candidate == null) nextNotFound();
 
   const { title: parsedTitle, body } = candidate.latestVersion.content;
 
@@ -73,6 +92,19 @@ export async function generateMetadata({ params }) {
   };
 }
 
-export default function Page(props) {
-  return <CandidateScreenClientWrapper {...props} />;
+export default async function Page({ params }) {
+  const candidate = await fetchCandidate(parseId(params.id), {
+    chainId: getChainId(),
+  });
+
+  if (candidate == null) nextNotFound();
+
+  return (
+    <ClientAppProvider>
+      <CandidateScreen candidateId={candidate.id} />
+      <StoreHydrater
+        state={{ proposalCandidatesById: { [candidate.id]: candidate } }}
+      />
+    </ClientAppProvider>
+  );
 }